% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{col_create_index}
\alias{col_create_index}
\title{Create an index on a collection field}
\usage{
col_create_index(col, field_name, index_param, option = NULL)
}
\arguments{
\item{col}{A collection object.}

\item{field_name}{Character. Name of the field to index.}

\item{index_param}{An index param object (e.g., from [hnsw_index_param()]).}

\item{option}{Optional index creation options, or `NULL`.}
}
\value{
The result of the create_index call (invisibly).
}
\description{
Create an index on a collection field
}
