% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{col_delete}
\alias{col_delete}
\title{Delete documents by ID}
\usage{
col_delete(col, ids)
}
\arguments{
\item{col}{A collection object.}

\item{ids}{Character vector of document IDs to delete.}
}
\value{
The result of the delete call (invisibly).
}
\description{
Delete documents by ID
}
