% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{col_delete_by_filter}
\alias{col_delete_by_filter}
\title{Delete documents by filter}
\usage{
col_delete_by_filter(col, filter)
}
\arguments{
\item{col}{A collection object.}

\item{filter}{A filter expression.}
}
\value{
The result of the delete call (invisibly).
}
\description{
Delete documents by filter
}
