% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{col_drop_index}
\alias{col_drop_index}
\title{Drop an index from a collection field}
\usage{
col_drop_index(col, field_name)
}
\arguments{
\item{col}{A collection object.}

\item{field_name}{Character. Name of the field whose index to drop.}
}
\value{
The result of the drop_index call (invisibly).
}
\description{
Drop an index from a collection field
}
