% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{col_fetch}
\alias{col_fetch}
\title{Fetch documents by ID}
\usage{
col_fetch(col, ids)
}
\arguments{
\item{col}{A collection object.}

\item{ids}{Character vector of document IDs to fetch.}
}
\value{
A list of [zvec_doc()] objects.
}
\description{
Fetch documents by ID
}
