% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{col_flush}
\alias{col_flush}
\title{Flush a collection to disk}
\usage{
col_flush(col)
}
\arguments{
\item{col}{A collection object.}
}
\value{
Invisibly `NULL`.
}
\description{
Flush a collection to disk
}
