% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{col_insert}
\alias{col_insert}
\title{Insert documents into a collection}
\usage{
col_insert(col, ...)
}
\arguments{
\item{col}{A collection object returned by [create_collection()] or
[open_collection()].}

\item{...}{[zvec_doc()] objects to insert.}
}
\value{
The result of the insert call (invisibly).
}
\description{
Insert documents into a collection
}
