% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{col_schema}
\alias{col_schema}
\title{Get the schema of a collection}
\usage{
col_schema(col)
}
\arguments{
\item{col}{A collection object.}
}
\value{
The Python `CollectionSchema` of the collection.
}
\description{
Get the schema of a collection
}
