% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{col_stats}
\alias{col_stats}
\title{Get statistics for a collection}
\usage{
col_stats(col)
}
\arguments{
\item{col}{A collection object.}
}
\value{
A statistics object for the collection.
}
\description{
Get statistics for a collection
}
