% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{col_update}
\alias{col_update}
\title{Update documents in a collection}
\usage{
col_update(col, ...)
}
\arguments{
\item{col}{A collection object.}

\item{...}{[zvec_doc()] objects to update.}
}
\value{
The result of the update call (invisibly).
}
\description{
Update documents in a collection
}
