% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{col_upsert}
\alias{col_upsert}
\title{Upsert documents into a collection}
\usage{
col_upsert(col, ...)
}
\arguments{
\item{col}{A collection object.}

\item{...}{[zvec_doc()] objects to upsert.}
}
\value{
The result of the upsert call (invisibly).
}
\description{
Upsert documents into a collection
}
