% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{collection_schema}
\alias{collection_schema}
\title{Create a collection schema}
\usage{
collection_schema(name, fields = NULL, vectors = NULL)
}
\arguments{
\item{name}{Character. Name of the collection.}

\item{fields}{List of field schemas created with [field_schema()], or `NULL`.}

\item{vectors}{A single [vector_schema()] or list thereof, or `NULL`.}
}
\value{
A Python `CollectionSchema` object.
}
\description{
Create a collection schema
}
