% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{field_schema}
\alias{field_schema}
\title{Create a field schema}
\usage{
field_schema(name, data_type, nullable = FALSE, index_param = NULL)
}
\arguments{
\item{name}{Character. Field name.}

\item{data_type}{A data type constant from [zvec_data_type()].}

\item{nullable}{Logical. Whether the field is nullable. Default `FALSE`.}

\item{index_param}{An index param object, or `NULL`.}
}
\value{
A Python `FieldSchema` object.
}
\description{
Create a field schema
}
