% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{flat_index_param}
\alias{flat_index_param}
\title{Create flat index parameters}
\usage{
flat_index_param(...)
}
\arguments{
\item{...}{Named arguments passed to `zvec$FlatIndexParam()`.}
}
\value{
A Python `FlatIndexParam` object.
}
\description{
Create flat index parameters
}
