% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{hnsw_index_param}
\alias{hnsw_index_param}
\title{Create HNSW index parameters}
\usage{
hnsw_index_param(...)
}
\arguments{
\item{...}{Named arguments passed to `zvec$HnswIndexParam()`.}
}
\value{
A Python `HnswIndexParam` object.
}
\description{
Create HNSW index parameters
}
