% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{invert_index_param}
\alias{invert_index_param}
\title{Create inverted index parameters}
\usage{
invert_index_param(...)
}
\arguments{
\item{...}{Named arguments passed to `zvec$InvertIndexParam()`.}
}
\value{
A Python `InvertIndexParam` object.
}
\description{
Create inverted index parameters
}
