% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{ivf_index_param}
\alias{ivf_index_param}
\title{Create IVF index parameters}
\usage{
ivf_index_param(...)
}
\arguments{
\item{...}{Named arguments passed to `zvec$IVFIndexParam()`.}
}
\value{
A Python `IVFIndexParam` object.
}
\description{
Create IVF index parameters
}
