% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{open_collection}
\alias{open_collection}
\title{Open an existing collection}
\usage{
open_collection(path, option = NULL)
}
\arguments{
\item{path}{Character. Directory path of the collection.}

\item{option}{Optional collection options, or `NULL`.}
}
\value{
A Python collection object.
}
\description{
Open an existing collection
}
