% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{vector_query}
\alias{vector_query}
\title{Create a vector query}
\usage{
vector_query(field_name, vector = NULL, id = NULL, param = NULL)
}
\arguments{
\item{field_name}{Character. The name of the vector field to search.}

\item{vector}{Numeric vector of query values, or `NULL`.}

\item{id}{Character document ID for ID-based lookup, or `NULL`.}

\item{param}{Additional query parameters, or `NULL`.}
}
\value{
A Python `VectorQuery` object.
}
\description{
Create a vector query
}
