% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{vector_schema}
\alias{vector_schema}
\title{Create a vector field schema}
\usage{
vector_schema(name, data_type, dimension, index_param = NULL)
}
\arguments{
\item{name}{Character. Vector field name.}

\item{data_type}{A data type constant from [zvec_data_type()], typically
`zvec_data_type()$VECTOR_FP32`.}

\item{dimension}{Integer. Number of dimensions.}

\item{index_param}{An index param object, or `NULL`.}
}
\value{
A Python `VectorSchema` object.
}
\description{
Create a vector field schema
}
